#23 [RENAME] input 컴포넌트를 component/Input 폴더 하위로 이동
