+++ src/pages/FindAccount/FindId.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Input from '../../components/Input';
+import Input from '../../components/Input/Input';
 
 export default function FindId() {
   return (
