+++ src/pages/SignUp/SignUpInfo.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import submitSignUpInfo from '../../apis/signUp';
-import BirthDayInput from '../../components/BirthDayInput';
 import OrangeButton from '../../components/Button/OrangeButton';
-import DuplicateCheckInput from '../../components/DuplicateCheckInput';
 import FieldSet from '../../components/FieldSet';
-import Input from '../../components/Input';
+import BirthDayInput from '../../components/Input/BirthDayInput';
+import DuplicateCheckInput from '../../components/Input/DuplicateCheckInput';
+import Input from '../../components/Input/Input';
 import Label from '../../components/Label';
 import ResetIcon from '../../components/ResetIcon';
 import ValidationMessage from '../../components/ValidationMessage';
